Add files via upload
This repository contains a simple console application built with C++ and named as the 'Railway Reservation System' project. This program features two modes, password protection and some basic file handling operations to add and display train inquiry and reservation records. This project is complete and totally-error free, and I have presented the source code in a very understandable manner.